+++ inst/examples/ex-plot_ford.R
@@ -1,0 +1,6 @@
+## Ceramic data
+## (data from Desachy 2004)
+data("compiegne", package = "folio")
+
+## Plot a Ford diagram
+plot_ford(compiegne)